Added check role when get user details, documents

+++ controllers/ClientsController.php
@@ -91,6 +91,10 @@ class ClientsController extends Controller
     }
 
     public function actionIndex() {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $service = new UserService();
 
         $search = Yii::$app->request->post('search');
@@ -122,6 +126,10 @@ class ClientsController extends Controller
     }
 
     public function actionDetails($clientId) {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $orderService = new OrdersService();
         $user = $this->clientsService->getClientDetails($clientId);
         $dataProvider = new ActiveDataProvider([
@@ -142,6 +150,10 @@ class ClientsController extends Controller
     }
 
     public function actionDetailsDocuments($clientId) {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $subclients = array();
         $documentsCalendar = array();
         $user = $this->clientsService->getClientDetails($clientId);
@@ -161,6 +173,10 @@ class ClientsController extends Controller
 
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         //add user can delete
         $userService = new UserService();
         $userService->deleteClient($id);
@@ -169,6 +185,10 @@ class ClientsController extends Controller
 
     public function actionDocuments()
     {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $currentUserId = Yii::$app->user->getId();
         $user = User::findOne($currentUserId);
         $subclients = array();
@@ -246,6 +266,10 @@ class ClientsController extends Controller
     }
 
     public function actionApproveOrder() {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $this->enableCsrfValidation = false;
         Yii::$app->response->format = Response::FORMAT_JSON;
 
@@ -263,6 +287,10 @@ class ClientsController extends Controller
     }
 
     public function actionDeleteOrder() {
+        if (!Yii::$app->user->can('employee')) {
+            return $this->goHome();
+        }
+
         $this->enableCsrfValidation = false;
         Yii::$app->response->format = Response::FORMAT_JSON;
